+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Bugs Fixed
 
 - **#2091** The variant panel for the `7` key was the variant panel for `4`.
+- **#2093** Inline shortcuts can be corrected with backspace, i.e. typing 
+- `sen[backspace][backspace]in` will be corrected to `\\sin`.
 
 ## 0.95.4 (2023-08-11)